image/requirements.txt
fastapi==0.55.1
handright==5.5.0
Pillow==7.1.2
qcloud_cos==3.3.6
